Move createBugsnagClient to exception handling service and fix tests
